requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-ai-formrecognizer==3.3.1
azure-ai-translation-text==1.0.1
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.14.1
azure-search==1.0.0b2
azure-search-documents==11.4.0b11
azure-storage-blob==12.20.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
deepdiff==8.0.1
distro==1.9.0
docx2txt==0.8
docx2txt2==1.0.4
emoji==2.14.0
entrypoints==0.4
et-xmlfile==1.1.0
faiss-cpu==1.9.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.8
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.81
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.22.0
msal==1.31.0
msal-extensions==1.2.0
msgspec==0.18.6
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.35.7
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
Pillow==10.1.0
portalocker==2.10.1
propcache==0.2.0
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
PyYAML==6.0.2
RapidFuzz==3.10.0
redis==5.1.1
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
toolz==1.0.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.15.14
unstructured-client==0.25.9
urllib3==2.2.3
uvicorn==0.31.1
waitress==3.0.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.15.2
